Implement transport protocol class
